Move all code under src/ and just re-export from index.ts
