+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "build/index.js",
   "scripts": {
-    "test": "jest",
+    "test": "tsc && jest",
     "prepare": "tsc"
   },
   "author": "",
